\name{alphaStart}
\alias{alphaStart}
\title{Identify item names that do not begin with a alphabetic character}
\usage{
  alphaStart(OCstack)
}
\arguments{
  \item{OCstack}{A stacked dataset of the CRFs see
  read.forms}
}
\description{
  Identify item names that do not begin with a alphabetic
  character
}
\author{
  Tyler Hunt \email{tyler.hunt@utah.edu}
}

