\name{item.length}
\alias{item.length}
\title{Test the length of the name of items.}
\usage{
  item.length(OCstack, length = 32, type = "item")
}
\arguments{
  \item{OCstack}{A stacked dataset of the CRFs see
  read.forms}

  \item{length}{The maximum item name length desired}

  \item{type}{can be item, description, or label.}
}
\description{
  Allows for users to test the length of items so they can
  be shortened if neccessary.
}
\author{
  Tyler Hunt \email{tyler.hunt@utah.edu}
}

