\name{read.forms}
\alias{read.forms}
\title{Read in OC CRF excel files}
\usage{
  read.forms(path)
}
\arguments{
  \item{path}{the path of the folder that the CRFs are in.}
}
\description{
  This function reads in and stacks the items sheet of the
  OC CRFs
}
\author{
  Tyler Hunt \email{tyler.hunt@utah.edu}
}

