\name{specialChars}
\alias{specialChars}
\title{Identify item names with special characters}
\usage{
  specialChars(OCstack)
}
\arguments{
  \item{OCstack}{A stacked dataset of the CRFs see
  read.forms}
}
\description{
  Identify item names with special characters
}
\author{
  Tyler Hunt \email{tyler.hunt@utah.edu}
}

