\name{underscores}
\alias{underscores}
\title{Identify item names with excessive underscores}
\usage{
  underscores(OCstack, allowed = 2)
}
\arguments{
  \item{OCstack}{A stacked dataset of the CRFs see
  read.forms}

  \item{allowed}{the number of underscores allowed in an
  item name}
}
\description{
  Identify item names with excessive underscores
}
\author{
  Tyler Hunt \email{tyler.hunt@utah.edu}
}

